fix: 회원가입 요청 url 변경
- POST /users -> POST /user/register
